+++ client/src/live-search/index.js
@@ -1,0 +1,1 @@
+export { default as LiveSearchInput } from './LiveSearchInput';
