+++ client/src/live-search/styled.js
@@ -1,5 +1,17 @@
 import Dropdown from 'react-bootstrap/Dropdown';
+import FormControl from 'react-bootstrap/FormControl';
 import styled from 'styled-components';
+
+export const StyledInputContainer = styled.div`
+  width: 600px;
+  margin-top: 10px;
+`;
+
+export const StyledFormControl = styled(FormControl)`
+  &:focus {
+    box-shadow: none;
+  }
+`;
 
 export const StyledDropdown = styled(Dropdown)`
   position: fixed;
