+++ client/src/search-box/index.js
@@ -1,0 +1,1 @@
+export { default as SearchBox } from './SearchBox';
