+++ client/src/search-box/styled.js
@@ -1,5 +1,4 @@
 import Card from 'react-bootstrap/Card';
-import FormControl from 'react-bootstrap/FormControl';
 import styled from 'styled-components';
 
 export const SearchBoxContainer = styled.div`
@@ -8,11 +7,6 @@ export const SearchBoxContainer = styled.div`
   flex-direction: column;
   align-items: center;
   justify-content: center;
-`;
-
-export const SearchInputContainer = styled.div`
-  width: 600px;
-  margin-top: 10px;
 `;
 
 export const SearchCardHeader = styled(Card.Header)`
@@ -29,10 +23,4 @@ export const SearchCardBody = styled(Card.Body)`
   width: 800px;
   height: 300px;
   padding-bottom: 70px;
-`;
-
-export const SearchFormControl = styled(FormControl)`
-  &:focus {
-    box-shadow: none;
-  }
 `;
